Now goes to soonest-due non-complete assignment
Increases ind variable if the assignment corresponding to the ind is complete, but make it to where the ind is still corresponding to an actual assignment so it still eases convenience even if all assignments are completed. A lot of lines are for checking if things exists just to make sure the content script does not crash.