Deleted unnecessary dropbox link
In the assignment detail page, there is dropbox link at the bottom of the page, all the way from the very beginning of development. Since there is a better-looking dropbox link on the top right of that page now, the one on the bottom is unnecessary and has thus been deleted.